error check if jail created

+++ plexinstallplex.sh
@@ -86,9 +86,13 @@ fi
 #
 # Create Jail
 echo '{"pkgs":["nano","ca_root_nss"]}' > /tmp/pkg.json
-iocage create --name "${JAIL_NAME}" -p /tmp/pkg.json -r ${RELEASE} ip4_addr="${INTERFACE}|${JAIL_IP}/24" defaultrouter="${DEFAULT_GW_IP}" boot="on" host_hostname="${JAIL_NAME}" vnet="${VNET}"
-
+if ! iocage create --name "${JAIL_NAME}" -p /tmp/pkg.json -r "${RELEASE}" ip4_addr="${INTERFACE}|${JAIL_IP}/24" defaultrouter="${DEFAULT_GW_IP}" boot="on" host_hostname="${JAIL_NAME}" vnet="${VNET}"
+then
+	echo "Failed to create jail"
+	exit 1
+fi
 rm /tmp/pkg.json
+
 echo ${POOL_PATH}/${APPS_PATH}/${PLEX_DATA}
 mkdir -p ${POOL_PATH}/${APPS_PATH}/${PLEX_DATA}
 mkdir -p ${POOL_PATH}/${MEDIA_LOCATION}
